Show item label in item editor header
